test(a11y): promote the no-injected-<style> guard from praxisCard/ to frontend/src
Same shape as the no-raw-style-values ratchet: a grandfathered allowlist that
only shrinks. LevelUpPopup's reduced-motion-guarded confetti keyframe is the one
grandfathered entry (no index.css home yet, out of this issue's scope). A second
assertion keeps the list honest — every entry must still exist and still inject.
